requirements.txt
Django==1.9.4
django-forms-bootstrap==3.0.1
django-taggit==0.18.0
Pillow==3.1.1
